feat(core, reader, ui): add intro chapters

+++ platform-reader/src/main/java/com/youversion/platform/reader/BibleReader.kt
@@ -110,7 +110,7 @@ fun BibleReader(
                                 BibleReference(
                                     versionId = versionId,
                                     bookUSFM = bookCode,
-                                    chapter = chapter.toInt(),
+                                    chapter = chapter,
                                 ).also { bibleReaderViewModel.onHeaderSelectionChange(it) }
                                 navController.popBackStack()
                             },
